feat: moving all unit tests over to TS, fixing issues
