+++ __tests__/cmds/categories/create.test.ts
@@ -1,83 +1,12 @@
-const nock = require('nock');
+import nock from 'nock';
 
-const getApiNock = require('../get-api-nock');
+import CategoriesCreateCommand from '../../../src/cmds/categories/create';
+import getAPIMock, { getAPIMockWithVersionHeader } from '../../helpers/get-api-mock';
 
-const CategoriesCommand = require('../../src/cmds/categories');
-const CategoriesCreateCommand = require('../../src/cmds/categories/create');
-
-const categories = new CategoriesCommand();
 const categoriesCreate = new CategoriesCreateCommand();
 
 const key = 'API_KEY';
 const version = '1.0.0';
-
-function getNockWithVersionHeader(v) {
-  return getApiNock({
-    'x-readme-version': v,
-  });
-}
-
-describe('rdme categories', () => {
-  beforeAll(() => nock.disableNetConnect());
-
-  afterEach(() => nock.cleanAll());
-
-  it('should error if no api key provided', () => {
-    return expect(categories.run({})).rejects.toStrictEqual(
-      new Error('No project API key provided. Please use `--key`.')
-    );
-  });
-
-  it('should return all categories for a single page', async () => {
-    const getMock = getNockWithVersionHeader(version)
-      .persist()
-      .get('/api/v1/categories?perPage=20&page=1')
-      .basicAuth({ user: key })
-      .reply(200, [{ title: 'One Category', slug: 'one-category', type: 'guide' }], {
-        'x-total-count': '1',
-      });
-
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
-
-    await expect(categories.run({ key, version: '1.0.0' })).resolves.toBe(
-      JSON.stringify([{ title: 'One Category', slug: 'one-category', type: 'guide' }], null, 2)
-    );
-
-    getMock.done();
-    versionMock.done();
-  });
-
-  it('should return all categories for multiple pages', async () => {
-    const getMock = getNockWithVersionHeader(version)
-      .persist()
-      .get('/api/v1/categories?perPage=20&page=1')
-      .basicAuth({ user: key })
-      .reply(200, [{ title: 'One Category', slug: 'one-category', type: 'guide' }], {
-        'x-total-count': '21',
-      })
-      .get('/api/v1/categories?perPage=20&page=2')
-      .basicAuth({ user: key })
-      .reply(200, [{ title: 'Another Category', slug: 'another-category', type: 'guide' }], {
-        'x-total-count': '21',
-      });
-
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
-
-    await expect(categories.run({ key, version: '1.0.0' })).resolves.toBe(
-      JSON.stringify(
-        [
-          { title: 'One Category', slug: 'one-category', type: 'guide' },
-          { title: 'Another Category', slug: 'another-category', type: 'guide' },
-        ],
-        null,
-        2
-      )
-    );
-
-    getMock.done();
-    versionMock.done();
-  });
-});
 
 describe('rdme categories:create', () => {
   beforeAll(() => nock.disableNetConnect());
@@ -104,12 +33,13 @@ describe('rdme categories:create', () => {
 
   it('should error if categoryType is not `guide` or `reference`', () => {
     return expect(
+      // @ts-expect-error Testing a CLI arg failure case.
       categoriesCreate.run({ key: '123', title: 'Test Title', categoryType: 'test' })
     ).rejects.toStrictEqual(new Error('`categoryType` must be `guide` or `reference`.'));
   });
 
   it('should create a new category if the title and type do not match and preventDuplicates=true', async () => {
-    const getMock = getNockWithVersionHeader(version)
+    const getMock = getAPIMockWithVersionHeader(version)
       .persist()
       .get('/api/v1/categories?perPage=20&page=1')
       .basicAuth({ user: key })
@@ -117,12 +47,12 @@ describe('rdme categories:create', () => {
         'x-total-count': '1',
       });
 
-    const postMock = getNockWithVersionHeader(version)
+    const postMock = getAPIMockWithVersionHeader(version)
       .post('/api/v1/categories')
       .basicAuth({ user: key })
       .reply(201, { title: 'New Category', slug: 'new-category', type: 'guide', id: '123' });
 
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
+    const versionMock = getAPIMock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
 
     await expect(
       categoriesCreate.run({
@@ -140,7 +70,7 @@ describe('rdme categories:create', () => {
   });
 
   it('should create a new category if the title matches but the type does not match and preventDuplicates=true', async () => {
-    const getMock = getNockWithVersionHeader(version)
+    const getMock = getAPIMockWithVersionHeader(version)
       .persist()
       .get('/api/v1/categories?perPage=20&page=1')
       .basicAuth({ user: key })
@@ -148,12 +78,12 @@ describe('rdme categories:create', () => {
         'x-total-count': '1',
       });
 
-    const postMock = getNockWithVersionHeader(version)
+    const postMock = getAPIMockWithVersionHeader(version)
       .post('/api/v1/categories')
       .basicAuth({ user: key })
       .reply(201, { title: 'Category', slug: 'category', type: 'reference', id: '123' });
 
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
+    const versionMock = getAPIMock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
 
     await expect(
       categoriesCreate.run({
@@ -171,12 +101,12 @@ describe('rdme categories:create', () => {
   });
 
   it('should create a new category if the title and type match and preventDuplicates=false', async () => {
-    const postMock = getNockWithVersionHeader(version)
+    const postMock = getAPIMockWithVersionHeader(version)
       .post('/api/v1/categories')
       .basicAuth({ user: key })
       .reply(201, { title: 'Category', slug: 'category', type: 'reference', id: '123' });
 
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
+    const versionMock = getAPIMock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
 
     await expect(
       categoriesCreate.run({
@@ -192,7 +122,7 @@ describe('rdme categories:create', () => {
   });
 
   it('should not create a new category if the title and type match and preventDuplicates=true', async () => {
-    const getMock = getNockWithVersionHeader(version)
+    const getMock = getAPIMockWithVersionHeader(version)
       .persist()
       .get('/api/v1/categories?perPage=20&page=1')
       .basicAuth({ user: key })
@@ -200,7 +130,7 @@ describe('rdme categories:create', () => {
         'x-total-count': '1',
       });
 
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
+    const versionMock = getAPIMock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
 
     await expect(
       categoriesCreate.run({
@@ -221,7 +151,7 @@ describe('rdme categories:create', () => {
   });
 
   it('should not create a new category if the non case sensitive title and type match and preventDuplicates=true', async () => {
-    const getMock = getNockWithVersionHeader(version)
+    const getMock = getAPIMockWithVersionHeader(version)
       .persist()
       .get('/api/v1/categories?perPage=20&page=1')
       .basicAuth({ user: key })
@@ -229,7 +159,7 @@ describe('rdme categories:create', () => {
         'x-total-count': '1',
       });
 
-    const versionMock = getApiNock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
+    const versionMock = getAPIMock().get(`/api/v1/version/${version}`).basicAuth({ user: key }).reply(200, { version });
 
     await expect(
       categoriesCreate.run({
