+++ __tests__/cmds/custompages/index.test.ts
@@ -1,26 +1,20 @@
-const nock = require('nock');
-const chalk = require('chalk');
-const fs = require('fs');
-const path = require('path');
-const crypto = require('crypto');
-const frontMatter = require('gray-matter');
+import fs from 'fs';
+import path from 'path';
 
-const APIError = require('../../src/lib/apiError');
-const getApiNock = require('../get-api-nock');
+import chalk from 'chalk';
+import frontMatter from 'gray-matter';
+import nock from 'nock';
 
-const CustomPagesCommand = require('../../src/cmds/custompages');
-const SingleCustomPageCommand = require('../../src/cmds/custompages/single');
+import CustomPagesCommand from '../../../src/cmds/custompages';
+import APIError from '../../../src/lib/apiError';
+import getAPIMock from '../../helpers/get-api-mock';
+import hashFileContents from '../../helpers/hash-file-contents';
 
 const custompages = new CustomPagesCommand();
-const customPagesSingle = new SingleCustomPageCommand();
 
 const fixturesBaseDir = '__fixtures__/custompages';
-const fullFixturesDir = `${__dirname}./../${fixturesBaseDir}`;
+const fullFixturesDir = `${__dirname}./../../${fixturesBaseDir}`;
 const key = 'API_KEY';
-
-function hashFileContents(contents) {
-  return crypto.createHash('sha1').update(contents).digest('hex');
-}
 
 describe('rdme custompages', () => {
   beforeAll(() => nock.disableNetConnect());
@@ -74,7 +68,7 @@ describe('rdme custompages', () => {
     it('should fetch custom page and merge with what is returned', () => {
       expect.assertions(1);
 
-      const getMocks = getApiNock()
+      const getMocks = getAPIMock()
         .get('/api/v1/custompages/simple-doc')
         .basicAuth({ user: key })
         .reply(200, { slug: simpleDoc.slug, htmlmode: false, lastUpdatedHash: 'anOldHash' })
@@ -82,7 +76,7 @@ describe('rdme custompages', () => {
         .basicAuth({ user: key })
         .reply(200, { slug: anotherDoc.slug, htmlmode: false, lastUpdatedHash: 'anOldHash' });
 
-      const updateMocks = getApiNock()
+      const updateMocks = getAPIMock()
         .put('/api/v1/custompages/simple-doc', {
           slug: simpleDoc.slug,
           body: simpleDoc.doc.content,
@@ -124,7 +118,7 @@ describe('rdme custompages', () => {
     it('should return custom page update info for dry run', () => {
       expect.assertions(1);
 
-      const getMocks = getApiNock()
+      const getMocks = getAPIMock()
         .get('/api/v1/custompages/simple-doc')
         .basicAuth({ user: key })
         .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: 'anOldHash' })
@@ -155,7 +149,7 @@ describe('rdme custompages', () => {
     it('should not send requests for custompages that have not changed', () => {
       expect.assertions(1);
 
-      const getMocks = getApiNock()
+      const getMocks = getAPIMock()
         .get('/api/v1/custompages/simple-doc')
         .basicAuth({ user: key })
         .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: simpleDoc.hash })
@@ -178,7 +172,7 @@ describe('rdme custompages', () => {
     it('should adjust "no changes" message if in dry run', () => {
       expect.assertions(1);
 
-      const getMocks = getApiNock()
+      const getMocks = getAPIMock()
         .get('/api/v1/custompages/simple-doc')
         .basicAuth({ user: key })
         .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: simpleDoc.hash })
@@ -207,7 +201,7 @@ describe('rdme custompages', () => {
       const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
       const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
 
-      const getMock = getApiNock()
+      const getMock = getAPIMock()
         .get(`/api/v1/custompages/${slug}`)
         .basicAuth({ user: key })
         .reply(404, {
@@ -217,7 +211,7 @@ describe('rdme custompages', () => {
           help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
         });
 
-      const postMock = getApiNock()
+      const postMock = getAPIMock()
         .post('/api/v1/custompages', { slug, body: doc.content, ...doc.data, lastUpdatedHash: hash })
         .basicAuth({ user: key })
         .reply(201, { slug, body: doc.content, ...doc.data, lastUpdatedHash: hash });
@@ -235,7 +229,7 @@ describe('rdme custompages', () => {
       const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs-html/${slug}.html`)));
       const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/new-docs-html/${slug}.html`)));
 
-      const getMock = getApiNock()
+      const getMock = getAPIMock()
         .get(`/api/v1/custompages/${slug}`)
         .basicAuth({ user: key })
         .reply(404, {
@@ -245,7 +239,7 @@ describe('rdme custompages', () => {
           help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
         });
 
-      const postMock = getApiNock()
+      const postMock = getAPIMock()
         .post('/api/v1/custompages', { slug, html: doc.content, htmlmode: true, ...doc.data, lastUpdatedHash: hash })
         .basicAuth({ user: key })
         .reply(201, { slug, html: doc.content, htmlmode: true, ...doc.data, lastUpdatedHash: hash });
@@ -262,7 +256,7 @@ describe('rdme custompages', () => {
       const slug = 'new-doc';
       const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
 
-      const getMock = getApiNock()
+      const getMock = getAPIMock()
         .get(`/api/v1/custompages/${slug}`)
         .basicAuth({ user: key })
         .reply(404, {
@@ -301,7 +295,7 @@ describe('rdme custompages', () => {
       const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/${folder}/${slug}.md`)));
       const hashTwo = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/${folder}/${slugTwo}.md`)));
 
-      const getMocks = getApiNock()
+      const getMocks = getAPIMock()
         .get(`/api/v1/custompages/${slug}`)
         .basicAuth({ user: key })
         .reply(404, {
@@ -319,7 +313,7 @@ describe('rdme custompages', () => {
           help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
         });
 
-      const postMocks = getApiNock()
+      const postMocks = getAPIMock()
         .post('/api/v1/custompages', {
           slug: slugTwo,
           body: docTwo.content,
@@ -360,7 +354,7 @@ describe('rdme custompages', () => {
       const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/slug-docs/${slug}.md`)));
       const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/slug-docs/${slug}.md`)));
 
-      const getMock = getApiNock()
+      const getMock = getAPIMock()
         .get(`/api/v1/custompages/${doc.data.slug}`)
         .basicAuth({ user: key })
         .reply(404, {
@@ -370,7 +364,7 @@ describe('rdme custompages', () => {
           help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
         });
 
-      const postMock = getApiNock()
+      const postMock = getAPIMock()
         .post('/api/v1/custompages', { slug, body: doc.content, ...doc.data, lastUpdatedHash: hash })
         .basicAuth({ user: key })
         .reply(201, { slug: doc.data.slug, body: doc.content, ...doc.data, lastUpdatedHash: hash });
@@ -381,330 +375,6 @@ describe('rdme custompages', () => {
 
       getMock.done();
       postMock.done();
-    });
-  });
-});
-
-describe('rdme custompages:single', () => {
-  beforeAll(() => nock.disableNetConnect());
-
-  afterAll(() => nock.cleanAll());
-
-  it('should error if no api key provided', () => {
-    return expect(customPagesSingle.run({})).rejects.toStrictEqual(
-      new Error('No project API key provided. Please use `--key`.')
-    );
-  });
-
-  it('should error if no file path provided', () => {
-    return expect(customPagesSingle.run({ key })).rejects.toStrictEqual(
-      new Error('No file path provided. Usage `rdme custompages:single <file> [options]`.')
-    );
-  });
-
-  it('should error if the argument is not a Markdown file', async () => {
-    await expect(customPagesSingle.run({ key, filePath: 'not-a-markdown-file' })).rejects.toStrictEqual(
-      new Error('The file path specified is not a Markdown or HTML file.')
-    );
-  });
-
-  it('should error if file path cannot be found', async () => {
-    await expect(
-      customPagesSingle.run({ key, version: '1.0.0', filePath: 'non-existent-file.markdown' })
-    ).rejects.toThrow('ENOENT: no such file or directory');
-  });
-
-  describe('new custompages', () => {
-    it('should create new custom page', async () => {
-      const slug = 'new-doc';
-      const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
-      const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
-
-      const getMock = getApiNock()
-        .get(`/api/v1/custompages/${slug}`)
-        .basicAuth({ user: key })
-        .reply(404, {
-          error: 'CUSTOMPAGE_NOTFOUND',
-          message: `The custom page with the slug '${slug}' couldn't be found`,
-          suggestion: '...a suggestion to resolve the issue...',
-          help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-        });
-
-      const postMock = getApiNock()
-        .post('/api/v1/custompages', { slug, body: doc.content, ...doc.data, lastUpdatedHash: hash })
-        .basicAuth({ user: key })
-        .reply(201, { slug, body: doc.content, ...doc.data });
-
-      await expect(
-        customPagesSingle.run({ filePath: `./__tests__/${fixturesBaseDir}/new-docs/new-doc.md`, key })
-      ).resolves.toBe(
-        `🌱 successfully created 'new-doc' with contents from ./__tests__/${fixturesBaseDir}/new-docs/new-doc.md`
-      );
-
-      getMock.done();
-      postMock.done();
-    });
-
-    it('should create new HTML custom page', async () => {
-      const slug = 'new-doc';
-      const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs-html/${slug}.html`)));
-      const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/new-docs-html/${slug}.html`)));
-
-      const getMock = getApiNock()
-        .get(`/api/v1/custompages/${slug}`)
-        .basicAuth({ user: key })
-        .reply(404, {
-          error: 'CUSTOMPAGE_NOTFOUND',
-          message: `The custom page with the slug '${slug}' couldn't be found`,
-          suggestion: '...a suggestion to resolve the issue...',
-          help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-        });
-
-      const postMock = getApiNock()
-        .post('/api/v1/custompages', { slug, html: doc.content, htmlmode: true, ...doc.data, lastUpdatedHash: hash })
-        .basicAuth({ user: key })
-        .reply(201, { slug, html: doc.content, htmlmode: true, ...doc.data });
-
-      await expect(
-        customPagesSingle.run({ filePath: `./__tests__/${fixturesBaseDir}/new-docs-html/new-doc.html`, key })
-      ).resolves.toBe(
-        `🌱 successfully created 'new-doc' with contents from ./__tests__/${fixturesBaseDir}/new-docs-html/new-doc.html`
-      );
-
-      getMock.done();
-      postMock.done();
-    });
-
-    it('should return creation info for dry run', async () => {
-      const slug = 'new-doc';
-      const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/new-docs/${slug}.md`)));
-
-      const getMock = getApiNock()
-        .get(`/api/v1/custompages/${slug}`)
-        .basicAuth({ user: key })
-        .reply(404, {
-          error: 'CUSTOMPAGE_NOTFOUND',
-          message: `The custom page with the slug '${slug}' couldn't be found`,
-          suggestion: '...a suggestion to resolve the issue...',
-          help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-        });
-
-      await expect(
-        customPagesSingle.run({ dryRun: true, filePath: `./__tests__/${fixturesBaseDir}/new-docs/new-doc.md`, key })
-      ).resolves.toBe(
-        `🎭 dry run! This will create 'new-doc' with contents from ./__tests__/${fixturesBaseDir}/new-docs/new-doc.md with the following metadata: ${JSON.stringify(
-          doc.data
-        )}`
-      );
-
-      getMock.done();
-    });
-
-    it('should fail if the custom page is invalid', async () => {
-      const folder = 'failure-docs';
-      const slug = 'fail-doc';
-
-      const errorObject = {
-        error: 'CUSTOMPAGE_INVALID',
-        message: "We couldn't save this page (Custom page title cannot be blank).",
-        suggestion: 'Make sure all the data is correct, and the body is valid Markdown or HTML.',
-        docs: 'fake-metrics-uuid',
-        help: "If you need help, email support@readme.io and include the following link to your API log: 'fake-metrics-uuid'.",
-      };
-
-      const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/${folder}/${slug}.md`)));
-
-      const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/${folder}/${slug}.md`)));
-
-      const getMock = getApiNock()
-        .get(`/api/v1/custompages/${slug}`)
-        .basicAuth({ user: key })
-        .reply(404, {
-          error: 'CUSTOMPAGE_NOTFOUND',
-          message: `The custom page with the slug '${slug}' couldn't be found`,
-          suggestion: '...a suggestion to resolve the issue...',
-          help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-        });
-
-      const postMock = getApiNock()
-        .post('/api/v1/custompages', { slug, body: doc.content, htmlmode: false, ...doc.data, lastUpdatedHash: hash })
-        .basicAuth({ user: key })
-        .reply(400, errorObject);
-
-      const filePath = `./__tests__/${fixturesBaseDir}/failure-docs/fail-doc.md`;
-
-      const formattedErrorObject = {
-        ...errorObject,
-        message: `Error uploading ${chalk.underline(`${filePath}`)}:\n\n${errorObject.message}`,
-      };
-
-      await expect(customPagesSingle.run({ filePath: `${filePath}`, key })).rejects.toStrictEqual(
-        new APIError(formattedErrorObject)
-      );
-
-      getMock.done();
-      postMock.done();
-    });
-
-    it('should fail if some other error when retrieving page slug', async () => {
-      const slug = 'fail-doc';
-
-      const errorObject = {
-        error: 'INTERNAL_ERROR',
-        message: 'Unknown error (yikes)',
-        suggestion: '...a suggestion to resolve the issue...',
-        help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-      };
-
-      const getMock = getApiNock().get(`/api/v1/custompages/${slug}`).basicAuth({ user: key }).reply(500, errorObject);
-
-      const filePath = `./__tests__/${fixturesBaseDir}/failure-docs/fail-doc.md`;
-
-      const formattedErrorObject = {
-        ...errorObject,
-        message: `Error uploading ${chalk.underline(`${filePath}`)}:\n\n${errorObject.message}`,
-      };
-
-      await expect(customPagesSingle.run({ filePath: `${filePath}`, key })).rejects.toStrictEqual(
-        new APIError(formattedErrorObject)
-      );
-
-      getMock.done();
-    });
-  });
-
-  describe('slug metadata', () => {
-    it('should use provided slug', async () => {
-      const slug = 'new-doc-slug';
-      const doc = frontMatter(fs.readFileSync(path.join(fullFixturesDir, `/slug-docs/${slug}.md`)));
-      const hash = hashFileContents(fs.readFileSync(path.join(fullFixturesDir, `/slug-docs/${slug}.md`)));
-
-      const getMock = getApiNock()
-        .get(`/api/v1/custompages/${doc.data.slug}`)
-        .basicAuth({ user: key })
-        .reply(404, {
-          error: 'CUSTOMPAGE_NOTFOUND',
-          message: `The custom page with the slug '${slug}' couldn't be found`,
-          suggestion: '...a suggestion to resolve the issue...',
-          help: 'If you need help, email support@readme.io and mention log "fake-metrics-uuid".',
-        });
-
-      const postMock = getApiNock()
-        .post('/api/v1/custompages', { slug, body: doc.content, ...doc.data, lastUpdatedHash: hash })
-        .basicAuth({ user: key })
-        .reply(201, { slug: doc.data.slug, body: doc.content, ...doc.data, lastUpdatedHash: hash });
-
-      await expect(
-        customPagesSingle.run({ filePath: `./__tests__/${fixturesBaseDir}/slug-docs/new-doc-slug.md`, key })
-      ).resolves.toBe(
-        `🌱 successfully created 'marc-actually-wrote-a-test' with contents from ./__tests__/${fixturesBaseDir}/slug-docs/new-doc-slug.md`
-      );
-
-      getMock.done();
-      postMock.done();
-    });
-  });
-
-  describe('existing custompages', () => {
-    let simpleDoc;
-
-    beforeEach(() => {
-      const fileContents = fs.readFileSync(path.join(fullFixturesDir, '/existing-docs/simple-doc.md'));
-      simpleDoc = {
-        slug: 'simple-doc',
-        doc: frontMatter(fileContents),
-        hash: hashFileContents(fileContents),
-      };
-    });
-
-    it('should fetch custom page and merge with what is returned', () => {
-      const getMock = getApiNock()
-        .get('/api/v1/custompages/simple-doc')
-        .basicAuth({ user: key })
-        .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: 'anOldHash' });
-
-      const updateMock = getApiNock()
-        .put('/api/v1/custompages/simple-doc', {
-          slug: simpleDoc.slug,
-          body: simpleDoc.doc.content,
-          htmlmode: false,
-          lastUpdatedHash: simpleDoc.hash,
-          ...simpleDoc.doc.data,
-        })
-        .basicAuth({ user: key })
-        .reply(200, {
-          slug: simpleDoc.slug,
-          body: simpleDoc.doc.content,
-          htmlmode: false,
-        });
-
-      return customPagesSingle
-        .run({ filePath: `./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md`, key })
-        .then(updatedDocs => {
-          expect(updatedDocs).toBe(
-            `✏️ successfully updated 'simple-doc' with contents from ./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md`
-          );
-
-          getMock.done();
-          updateMock.done();
-        });
-    });
-
-    it('should return custom page update info for dry run', () => {
-      expect.assertions(1);
-
-      const getMock = getApiNock()
-        .get('/api/v1/custompages/simple-doc')
-        .basicAuth({ user: key })
-        .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: 'anOldHash' });
-
-      return customPagesSingle
-        .run({ dryRun: true, filePath: `./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md`, key })
-        .then(updatedDocs => {
-          // All custompages should have been updated because their hashes from the GET request were different from what they
-          // are currently.
-          expect(updatedDocs).toBe(
-            [
-              `🎭 dry run! This will update 'simple-doc' with contents from ./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md with the following metadata: ${JSON.stringify(
-                simpleDoc.doc.data
-              )}`,
-            ].join('\n')
-          );
-
-          getMock.done();
-        });
-    });
-
-    it('should not send requests for custompages that have not changed', () => {
-      expect.assertions(1);
-
-      const getMock = getApiNock()
-        .get('/api/v1/custompages/simple-doc')
-        .basicAuth({ user: key })
-        .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: simpleDoc.hash });
-
-      return customPagesSingle
-        .run({ filePath: `./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md`, key })
-        .then(skippedDocs => {
-          expect(skippedDocs).toBe('`simple-doc` was not updated because there were no changes.');
-
-          getMock.done();
-        });
-    });
-
-    it('should adjust "no changes" message if in dry run', () => {
-      const getMock = getApiNock()
-        .get('/api/v1/custompages/simple-doc')
-        .basicAuth({ user: key })
-        .reply(200, { slug: simpleDoc.slug, lastUpdatedHash: simpleDoc.hash });
-
-      return customPagesSingle
-        .run({ dryRun: true, filePath: `./__tests__/${fixturesBaseDir}/existing-docs/simple-doc.md`, key })
-        .then(skippedDocs => {
-          expect(skippedDocs).toBe('🎭 dry run! `simple-doc` will not be updated because there were no changes.');
-
-          getMock.done();
-        });
     });
   });
 });
