+++ __tests__/helpers/get-api-mock.ts
@@ -1,0 +1,19 @@
+import config from 'config';
+import nock from 'nock';
+
+import { getUserAgent } from '../../src/lib/fetch';
+
+export default function getAPIMock(reqHeaders = {}) {
+  return nock(config.get('host'), {
+    reqheaders: {
+      'User-Agent': getUserAgent(),
+      ...reqHeaders,
+    },
+  });
+}
+
+export function getAPIMockWithVersionHeader(v) {
+  return getAPIMock({
+    'x-readme-version': v,
+  });
+}
