+++ __tests__/helpers/hash-file-contents.ts
@@ -1,0 +1,5 @@
+import crypto from 'crypto';
+
+export default function hashFileContents(contents) {
+  return crypto.createHash('sha1').update(contents).digest('hex');
+}
