+++ package.json
@@ -76,6 +76,7 @@
     "alex": "^10.0.0",
     "eslint": "^8.21.0",
     "jest": "^28.1.1",
+    "jest-extended": "^3.0.2",
     "nock": "^13.2.7",
     "prettier": "^2.7.1",
     "ts-jest": "^28.0.7",
