+++ src/cmds/versions/delete.ts
@@ -33,9 +33,13 @@ export default class DeleteVersionCommand extends Command {
   }
 
   async run(opts: CommandOptions<{}>) {
-    super.run(opts, true);
+    super.run(opts);
 
     const { key, version } = opts;
+
+    if (!opts.key) {
+      return Promise.reject(new Error('No project API key provided. Please use `--key`.'));
+    }
 
     const selectedVersion = await getProjectVersion(version, key, false).catch(e => {
       return Promise.reject(e);
